Garmin: fix major issue with field encoding/decoding
Our implementation of scale and offset was backwards: we were adding offset and then dividing by scale when decoding fields, but the publicly available protocol description dictates otherwise ( http://web.archive.org/web/20240519102659/https://developer.garmin.com/fit/protocol/#scaleoffset ):  "the binary quantity is divided by the scale factor and then the offset is subtracted".
For this reason the sign of GARMIN_TIME_EPOCH in Timestamp field definition must be flipped as well.

+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/service/devices/garmin/fit/baseTypes/BaseType.java
@@ -48,11 +48,11 @@ public enum BaseType {
         return identifier;
     }
 
-    public Object decode(ByteBuffer byteBuffer, int scale, int offset) {
+    public Object decode(ByteBuffer byteBuffer, double scale, int offset) {
         return baseTypeInterface.decode(byteBuffer, scale, offset);
     }
 
-    public void encode(ByteBuffer byteBuffer, Object o, int scale, int offset) {
+    public void encode(ByteBuffer byteBuffer, Object o, double scale, int offset) {
         baseTypeInterface.encode(byteBuffer, o, scale, offset);
     }
