+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/service/devices/garmin/fit/baseTypes/BaseTypeInterface.java
@@ -5,9 +5,9 @@ import java.nio.ByteBuffer;
 public interface BaseTypeInterface {
     int getByteSize();
 
-    Object decode(ByteBuffer byteBuffer, int scale, int offset);
+    Object decode(ByteBuffer byteBuffer, double scale, int offset);
 
-    void encode(ByteBuffer byteBuffer, Object o, int scale, int offset);
+    void encode(ByteBuffer byteBuffer, Object o, double scale, int offset);
 
     void invalidate(ByteBuffer byteBuffer);
 }
